Correct 'flags' type returned by ddsketch_info
The SQL function was declared to return flags as BIGINT, but the code
was using Int32GetDatum(). The struct declares it as int32, so adjust
the declaration to use INT and use Int32GetDatum() in C.

+++ sql/ddsketch--2.0.0-dev.sql
@@ -122,7 +122,7 @@ CREATE OR REPLACE FUNCTION ddsketch_union(p_sketch1 ddsketch, p_sketch2 ddsketch
     LANGUAGE C IMMUTABLE;
 
 
-CREATE OR REPLACE FUNCTION ddsketch_info(p_sketch ddsketch, out bytes bigint, out flags bigint, out alpha double precision, out count bigint, out zero_count bigint, out max_buckets int, out negative_buckets int, out positive_buckets int, out min_indexable double precision, out max_indexable double precision)
+CREATE OR REPLACE FUNCTION ddsketch_info(p_sketch ddsketch, out bytes bigint, out flags int, out alpha double precision, out count bigint, out zero_count bigint, out max_buckets int, out negative_buckets int, out positive_buckets int, out min_indexable double precision, out max_indexable double precision)
     RETURNS record
     AS 'ddsketch', 'ddsketch_sketch_info'
     LANGUAGE C IMMUTABLE STRICT PARALLEL SAFE;
